pipeline/requirements.txt
aws-cdk.core>=1.134.0
aws-cdk.assertions>=1.134.0
aws-cdk.aws-s3
aws-cdk.aws_codebuild
aws-cdk.aws_codecommit
aws-cdk.aws_codepipeline
aws-cdk.aws_codepipeline_actions